refactor utils, fix minor consistency issues
organizing to prepare for texture export functionality
