+++ src/store/stage-data/process-stage-polygon-file/ramToRawAddress.ts
@@ -1,4 +1,4 @@
-import Offsets from './StructOffsets';
+import Offsets from '../../../constants/StructOffsets';
 
 export default function ramToBinAddress(a: number) {
   return a - Offsets.RAM;
