+++ src/utils/images/getImageDimensions.ts
